Rename src directory to bdbot
