+++ bdbot/cogs/Errors.py
