+++ bdbot/misc/ADD_COMIC.md
