+++ bdbot/misc/comics_details.json
