+++ bdbot/misc/random-footers.txt
