+++ bdbot/misc/retired_comics.json
